requirements.txt
anaconda==0.0.1.1
bleach==5.0.1; python_version >= "3.7"
certifi==2022.6.15; python_version >= "3.7" and python_version < "4"
charset-normalizer==2.1.0; python_version >= "3.7" and python_version < "4" and python_full_version >= "3.6.0"
click==8.0.4; python_version >= "3.6"
colorama==0.4.5; python_version >= "3.6" and python_full_version < "3.0.0" and platform_system == "Windows" or python_version >= "3.6" and python_full_version >= "3.5.0" and platform_system == "Windows"
commonmark==0.9.1; python_full_version >= "3.6.3" and python_full_version < "4.0.0"
configparser==5.2.0; python_version >= "3.6"
cycler==0.11.0; python_version >= "3.7"
fonttools==4.34.4; python_version >= "3.7"
greenlet==1.1.2; python_version >= "3" and python_full_version < "3.0.0" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and (python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0") and (python_version >= "3.6" and python_full_version < "3.0.0" and python_version < "4.0" or python_version >= "3.6" and python_version < "4.0" and python_full_version >= "3.6.0") or python_version >= "3" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and (python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0") and python_full_version >= "3.5.0" and (python_version >= "3.6" and python_full_version < "3.0.0" and python_version < "4.0" or python_version >= "3.6" and python_version < "4.0" and python_full_version >= "3.6.0")
idna==3.3; python_version >= "3.7" and python_version < "4"
importlib-metadata==4.12.0; python_version >= "3.7"
jinja2==3.1.2; python_version >= "3.7"
kiwisolver==1.4.4; python_version >= "3.7"
laceworksdk==1.4.0
markupsafe==2.1.1; python_version >= "3.7"
matplotlib==3.5.3; python_version >= "3.7"
networkx==2.8.5; python_version >= "3.8"
numpy==1.23.1
packaging==21.3; python_version >= "3.7"
pandas==1.4.3; python_version >= "3.8"
pillow==9.2.0; python_version >= "3.7"
psycopg2-binary==2.9.3; python_version >= "3.6"
pygments==2.12.0; python_full_version >= "3.6.3" and python_full_version < "4.0.0" and python_version >= "3.6"
pyparsing==3.0.9; python_full_version >= "3.6.8" and python_version >= "3.7"
python-dateutil==2.8.2; python_version >= "3.8" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.8"
python-dotenv==0.20.0; python_version >= "3.5"
pytz==2022.2; python_version >= "3.8"
requests==2.28.1; python_version >= "3.7" and python_version < "4"
rich==12.5.1; python_full_version >= "3.6.3" and python_full_version < "4.0.0"
scipy==1.9.0; python_version >= "3.8" and python_version < "3.11"
setuptools-scm==6.4.2; python_version >= "3.7"
shellingham==1.5.0; python_version >= "3.6"
six==1.16.0; python_version >= "3.8" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.8"
sqlalchemy-utils==0.38.3; python_version >= "3.6" and python_version < "4.0"
sqlalchemy==1.4.40; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.6.0")
tomli==2.0.1; python_version >= "3.7"
typer==0.4.2; python_version >= "3.6"
typing-extensions==4.3.0; python_full_version >= "3.6.3" and python_full_version < "4.0.0" and python_version < "3.9" and python_version >= "3.7"
urllib3==1.26.11; python_version >= "3.7" and python_full_version < "3.0.0" and python_version < "4" or python_full_version >= "3.6.0" and python_version < "4" and python_version >= "3.7"
webencodings==0.5.1; python_version >= "3.7"
zipp==3.8.1; python_version >= "3.7" and python_version < "3.8"
